add alembic support for db migrations and add APP_ENV

+++ snapper/database.py
@@ -8,7 +8,7 @@ envs = get_envs()
 DATABASE_URL = f'mysql+aiomysql://{envs["DATABASE_USER"]}:{envs["DATABASE_PASSWORD"]}@localhost/{envs["DATABASE_NAME"]}'
 
 # Create async engine and session
-engine = create_async_engine(DATABASE_URL, echo=True)
+engine: AsyncEngine = create_async_engine(DATABASE_URL, echo=True)
 AsyncSessionLocal = sessionmaker(bind=engine, class_=AsyncSession, expire_on_commit=False)  # type: ignore
 
 
@@ -23,10 +23,14 @@ class Base(DeclarativeBase):
     pass
 
 
-async def init_db():
-    async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
-        await conn.run_sync(Base.metadata.create_all)
+async def drop_and_create_db():
+    """Completely reset the database. Should only be used in "dev"-mode.
+    Double checking in __main__ method and here as well.
+    """
+    if envs["APP_ENV"] == "dev":
+        async with engine.begin() as conn:
+            await conn.run_sync(Base.metadata.drop_all)
+            # await conn.run_sync(Base.metadata.create_all)
 
 
 class Clip(Base):
